Revert "Merge pull request #26266 from jmccartie/jm/bigint"
This reverts commit d7f55e987849c5a17f9d3152abd4fbacac08a509, reversing
changes made to 575212a1ba2e4d170985531402c6e3d27af5a0ea.

Conflicts:
	activerecord/CHANGELOG.md
	activerecord/lib/active_record/connection_adapters/abstract_mysql_adapter.rb
	activerecord/lib/active_record/errors.rb
	activerecord/lib/active_record/migration/compatibility.rb

+++ activerecord/lib/active_record/connection_adapters/abstract/schema_definitions.rb
@@ -71,7 +71,7 @@ module ActiveRecord
         polymorphic: false,
         index: true,
         foreign_key: false,
-        type: :bigint,
+        type: :integer,
         **options
       )
         @name = name
